разделяет модуль утилс на 2 разных
