+++ src/utils/render.js
@@ -1,13 +1,6 @@
-import {RenderPosition} from './constants.js';
-
-const getRandomBetween = (min, max) => {
-  return Math.floor(Math.random() * (max - min) + min);
-};
-
-const getRandomArrayItem = (array) => {
-  const randomIndex = getRandomBetween(0, array.length);
-
-  return array[randomIndex];
+const RenderPosition = {
+  AFTERBEGIN: `afterbegin`,
+  BEFOREEND: `beforeend`
 };
 
 const createElement = (template) => {
@@ -27,4 +20,4 @@ const renderHtmlPart = (container, template, place) => {
   }
 };
 
-export {getRandomBetween, getRandomArrayItem, createElement, renderHtmlPart};
+export {createElement, renderHtmlPart, RenderPosition};
